backend/requirements.txt
fastapi
uvicorn
transformers
torch
pdf2image
Pillow
python-multipart
olmocr
python-dotenv